feat: 그룹 수정, 높이 수정
