+++ client/src/feature/Main/MainTodo.jsx
@@ -7,12 +7,12 @@ import MainTodoList from "./MainTodoList";
 const MainTodo = ({ groupId }) => {
   const { isLoading, data } = useQueryTodoGroup({ groupId });
 
-  if (isLoading) return <StyledWrapper />;
+  if (isLoading || !data?.groupInfo) return <StyledWrapper />;
 
   const { groupInfo, members } = data;
   return (
     <StyledWrapper>
-      <GroupInfo title={groupInfo.todo_group_title} members={members} />
+      <GroupInfo title={groupInfo?.todo_group_title} members={members} />
       <MainTodoList startDate={dayjs()} groupId={groupId} />
     </StyledWrapper>
   );
@@ -22,10 +22,8 @@ export default MainTodo;
 
 const StyledWrapper = styled.div`
   width: 50%;
-  height: 100%;
   padding: 6.4rem 2.4rem 0 6.4rem;
   position: relative;
-  overflow: show;
 
   .group-info {
     margin-bottom: 4rem;
