+++ client/src/pages/LedgerPage.jsx
@@ -52,6 +52,7 @@ export default LedgerPage;
 const StyledWrapper = styled.div`
   height: 100%;
   width: 100%;
+  overflow: hidden;
   display: flex;
   flex-direction: column;
   position: relative;
